# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Group Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthStatus {
  authenticated: boolean;
  phone?: string | null;
  username?: string | null;
  firstName?: string | null;
}

export interface AuthResult {
  success: boolean;
  message: string;
  requiresPassword?: boolean | null;
  authenticated?: boolean | null;
}

export interface SendCodeBody {
  phone: string;
}

export interface VerifyCodeBody {
  phone: string;
  code: string;
  phoneCodeHash: string;
}

export interface VerifyPasswordBody {
  password: string;
}

export type TelegramGroupType =
  (typeof TelegramGroupType)[keyof typeof TelegramGroupType];

export const TelegramGroupType = {
  group: "group",
  channel: "channel",
  supergroup: "supergroup",
} as const;

export interface TelegramGroup {
  id: string;
  title: string;
  username?: string | null;
  membersCount?: number | null;
  description?: string | null;
  type: TelegramGroupType;
  isPublic: boolean;
  inviteLink?: string | null;
}

export interface GroupSearchResults {
  results: TelegramGroup[];
  total: number;
}

export type GroupStatus = (typeof GroupStatus)[keyof typeof GroupStatus];

export const GroupStatus = {
  saved: "saved",
  joined: "joined",
  pending: "pending",
  failed: "failed",
} as const;

export interface Group {
  id: number;
  telegramId: string;
  title: string;
  username?: string | null;
  membersCount?: number | null;
  type: string;
  status: GroupStatus;
  createdAt: string;
}

export interface GroupList {
  groups: Group[];
  total: number;
}

export interface SaveGroupBody {
  telegramId: string;
  title: string;
  username?: string | null;
  membersCount?: number | null;
  type: string;
}

export interface JoinGroupsBody {
  groupIds: number[];
  delaySeconds?: number;
}

export type CampaignStatus =
  (typeof CampaignStatus)[keyof typeof CampaignStatus];

export const CampaignStatus = {
  draft: "draft",
  active: "active",
  paused: "paused",
  completed: "completed",
} as const;

export type CampaignScheduleType =
  (typeof CampaignScheduleType)[keyof typeof CampaignScheduleType];

export const CampaignScheduleType = {
  once: "once",
  hourly: "hourly",
  every2h: "every2h",
  every4h: "every4h",
  every8h: "every8h",
  every12h: "every12h",
  daily: "daily",
  custom: "custom",
} as const;

export interface Campaign {
  id: number;
  name: string;
  message: string;
  status: CampaignStatus;
  scheduleType: CampaignScheduleType;
  intervalHours?: number | null;
  targetGroupIds: number[];
  sentCount: number;
  failCount: number;
  lastRunAt?: string | null;
  nextRunAt?: string | null;
  createdAt: string;
}

export interface CampaignList {
  campaigns: Campaign[];
  total: number;
}

export type CreateCampaignBodyScheduleType =
  (typeof CreateCampaignBodyScheduleType)[keyof typeof CreateCampaignBodyScheduleType];

export const CreateCampaignBodyScheduleType = {
  once: "once",
  hourly: "hourly",
  every2h: "every2h",
  every4h: "every4h",
  every8h: "every8h",
  every12h: "every12h",
  daily: "daily",
  custom: "custom",
} as const;

export interface CreateCampaignBody {
  name: string;
  message: string;
  scheduleType: CreateCampaignBodyScheduleType;
  intervalHours?: number | null;
  targetGroupIds: number[];
}

export interface UpdateCampaignBody {
  name?: string;
  message?: string;
  scheduleType?: string;
  intervalHours?: number | null;
  targetGroupIds?: number[];
  status?: string;
}

export type JobType = (typeof JobType)[keyof typeof JobType];

export const JobType = {
  join: "join",
  broadcast: "broadcast",
} as const;

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus];

export const JobStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface Job {
  id: number;
  type: JobType;
  status: JobStatus;
  targetId?: string | null;
  campaignId?: number | null;
  groupId?: number | null;
  message?: string | null;
  error?: string | null;
  createdAt: string;
  completedAt?: string | null;
}

export interface JobList {
  jobs: Job[];
  total: number;
}

export interface JobResult {
  success: boolean;
  message: string;
  jobIds?: number[] | null;
}

export interface DeleteResult {
  success: boolean;
  message: string;
}

export interface DashboardStats {
  totalGroups: number;
  joinedGroups: number;
  activeCampaigns: number;
  totalSent: number;
  totalFailed: number;
  recentJobs: Job[];
  authStatus: AuthStatus;
}

export type SearchGroupsParams = {
  query?: string;
  minMembers?: number;
  maxMembers?: number;
  limit?: number;
  groupType?: SearchGroupsGroupType;
};

export type SearchGroupsGroupType =
  (typeof SearchGroupsGroupType)[keyof typeof SearchGroupsGroupType];

export const SearchGroupsGroupType = {
  all: "all",
  group: "group",
  channel: "channel",
  supergroup: "supergroup",
} as const;

export type ListGroupsParams = {
  status?: ListGroupsStatus;
};

export type ListGroupsStatus =
  (typeof ListGroupsStatus)[keyof typeof ListGroupsStatus];

export const ListGroupsStatus = {
  all: "all",
  saved: "saved",
  joined: "joined",
  pending: "pending",
} as const;

export type ListJobsParams = {
  limit?: number;
};