# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Group Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthResult {
  success: boolean;
  message: string;
  requiresPassword?: boolean | null;
  authenticated?: boolean | null;
}