# lib/api-zod/src/generated/types/groupList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Group Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { Group } from "./group";

export interface GroupList {
  groups: Group[];
  total: number;
}