# lib/api-zod/src/generated/types/saveGroupBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Group Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface SaveGroupBody {
  telegramId: string;
  title: string;
  username?: string | null;
  membersCount?: number | null;
  type: string;
}