# lib/api-zod/src/generated/types/updateCampaignBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Group Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateCampaignBody {
  name?: string;
  message?: string;
  scheduleType?: string;
  intervalHours?: number | null;
  targetGroupIds?: number[];
  status?: string;
}