# lib/api-zod/src/generated/types/verifyCodeBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Group Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyCodeBody {
  phone: string;
  code: string;
  phoneCodeHash: string;
}